pyproject.toml
[project]
name = "devdz-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler==3.10.4",
    "python-dotenv>=1.1.0",
    "python-telegram-bot==20.7",
]
